refactor: extract RecordFormatter collaborator from RecordHoldersService
